Move Hashable (Tuple a) back to NestedTuple
Trying to keep the Hashable instances in a single
place creates orphan warnings so that was a bad
idea. Move them back to where they belong.
